# Fix #304 (simplify init clock)