--- a/rhine/src/FRP/Rhine/Clock/Realtime/Event.hs
+++ b/rhine/src/FRP/Rhine/Clock/Realtime/Event.hs
@@ -150,16 +150,11 @@ instance Semigroup (EventClock event) where
 instance (MonadIO m) => Clock (EventChanT event m) (EventClock event) where
   type Time (EventClock event) = UTCTime
   type Tag (EventClock event) = event
-  initClock _ = do
-    initialTime <- liftIO getCurrentTime
-    return
-      ( constM $ do
+  initClock _ = constM $ do
           chan <- ask
           event <- liftIO $ readChan chan
           time <- liftIO getCurrentTime
           return (time, event)
-      , initialTime
-      )
 
 instance GetClockProxy (EventClock event)