--- a/rhine/src/FRP/Rhine/Clock/Realtime/Never.hs
+++ b/rhine/src/FRP/Rhine/Clock/Realtime/Never.hs
@@ -27,11 +27,6 @@ instance (MonadIO m) => Clock m Never where
   type Time Never = UTCTime
   type Tag Never = Void
 
-  initClock _ = do
-    initialTime <- liftIO getCurrentTime
-    return
-      ( constM (liftIO . forever . threadDelay $ 10 ^ 9)
-      , initialTime
-      )
+  initClock _ = constM (liftIO . forever . threadDelay $ 10 ^ 9)
 
 instance GetClockProxy Never