--- a/rhine/src/FRP/Rhine/Clock/Unschedule.hs
+++ b/rhine/src/FRP/Rhine/Clock/Unschedule.hs
@@ -5,7 +5,6 @@
 module FRP.Rhine.Clock.Unschedule where
 
 -- base
-import Control.Arrow
 import Control.Concurrent qualified as Concurrent (yield)
 import Control.Monad.IO.Class
 
@@ -39,7 +38,7 @@ unyieldClock cl = UnscheduleClock cl $ const $ liftIO Concurrent.yield
 instance (TimeDomain (Time cl), Clock (ScheduleT (Diff (Time cl)) m) cl, Monad m) => Clock m (UnscheduleClock m cl) where
   type Tag (UnscheduleClock _ cl) = Tag cl
   type Time (UnscheduleClock _ cl) = Time cl
-  initClock UnscheduleClock {scheduleClock, scheduleWait} = run $ first (hoistS run) <$> initClock scheduleClock
+  initClock UnscheduleClock {scheduleClock, scheduleWait} = hoistS run $ initClock scheduleClock
     where
       run :: ScheduleT (Diff (Time cl)) m a -> m a
       run = runScheduleT scheduleWait