@turion 2024-08-13
This PR alone cannot give the speedups I'm hoping for, because clock erasure of `SN`s still doesn't happen. Only in combination with #348 I'll be able to get the full speedup I'm hoping for.