@turion 2024-08-13
#349 is helpful because it adds a more realistic benchmark where `initClock` is not inlined yet because it contains an `IO` action.